Homework-10 Bugfix-1
